+++ BOCCHI.Treasure/Services/ITreasureTracker.cs
@@ -6,6 +6,9 @@ public interface ITreasureTracker
 
     bool CountInitialised { get; }
 
+    /// <summary>UTC time of the last successful Treasure Sight WideText parse.</summary>
+    DateTime LastCountUpdateUtc { get; }
+
     int BronzeChests { get; }
 
     int SilverChests { get; }